lisp/org-table.el: Allow named columns on LHS
* lisp/org-table.el (org-table-recalculate): Add support for named
columns on the lhs of spreadsheet formulas, prioritizing named columns
over named fields if there is a conflict.
(org-table-edit-formulas): Modify category name to include column
formulas with field formulas.
(org-table-get-stored-formulas): Remove comment mentioning lack of named
columns on the lhs.
* testing/lisp/test-org-table.el (test-org-table/named-column): Add test
case for named columns.
* etc/ORG-NEWS (Spreadsheets now support named columns on the left-side
of a formula): Document the change.
* doc/org-manual.org (Advanced features): Document precedence of column
names over field names.
(Column formulas): Remove reference to disallowed column names on the lhs.

This change breaks compatibility in tables with a named field and named
column with the same name, when that name is used on the lhs of a formula.